cdk/chatHistory/requirements.txt
boto3
botocore
sqlalchemy
Pillow
pymupdf
psycopg[binary,pool]
psycopg2-binary
